Convert widget options to indexable form
Fixes #3505. Technically that issue is resolved by #3501, but the same
problem can happen for several other widgets that take options, so it is
worth applying a fix more broadly.

We already ensure the options are iterable, but this is not sufficient
for all of our needs, because we sometimes need to look up an index of a
value, which requires a `Sequence`. Additionally, users sometimes pass
in types that were not included in the existing list of supported types,
and we would like to more precisely define the requirements we have on
types and ensure our code can handle them.

Widgets now accept any `Collection` for an options parameter, because
the required iterability means we can fall back to turning it into a
list if it does not support the desired `index` operation. This should
work for nearly all types, though it is possible that some will require
special handling like Pandas DataFrames. (Do dataframes actually count
as `Collection`s? If not, the type annotation added here should be
extended to union with them)

To make mypy cooperate with some added type annotations, one of the
locals in the widgets has been renamed.