Convert widget options to indexable form (#3515)
* Convert widget options to indexable form

Fixes #3505. Technically that issue is resolved by #3501, but the same
problem can happen for several other widgets that take options, so it is
worth applying a fix more broadly.

We already ensure the options are iterable, but this is not sufficient
for all of our needs, because we sometimes need to look up an index of a
value, which requires a `Sequence`. Additionally, users sometimes pass
in types that were not included in the existing list of supported types,
and we would like to more precisely define the requirements we have on
types and ensure our code can handle them.

Widgets now accept any `Sequence`, or any of several other types that we
know how to convert to one. Other types are not necessarily supported,
but if converting it into a list gives the desired values it should work.

To make mypy cooperate with some added type annotations, one of the
locals in the widgets has been renamed.