Fix mypy errors by casting call to WhichOneOf (#3541)
It looks like the errors that didn't appear in the PR CI builds but
started appearing in the nightly and on develop are due to version 2.6
of mypy-protobuf being a bit more strict/correct with types. In this case,
we're guaranteed for the return value of WhichOneOf to be not None, so
we just need to cast this value.