Always open user-provided text files as utf8
This shouldn't matter at all for Mac and Linux users, but in Windows,
utf8 isn't the default encoding, so things may explode when the file
handler returned by a call to the Python builtin `open` stumbles upon a
Unicode character.

We chose to specify the encoding explicitly in calls to `open` rather
than other ways of doing this (such as via env var) since there are very
few calls to `open` that needed to be updated, anyway.