Make copies of returned reference types stored in widget state
#3600 is caused by the initial value of a widget, which is stored and
compared to internally, being returned directly as the widget's return
value. This is not a problem for value types, but if we return a mutable
reference type and the user mutates it later in the script, it changes
our stored copy, so we incorrectly think the initial value has changed
on the next run.

This is also a potential issue for values that are passed in, in various
ways, but handling them systematically will require some careful
checking. The reference type sharing introduces a potential for bugs
akin to "shotgun parsing", and Python is not very well equiped for
dealing with this kind of issue.

The unit test for this might make more sense elsewhere, but I couldn't
decide where.