Give UploadedFile metadata value equality
This is to make the file uploader e2e tests pass, because they relied on
reference equality of UploadedFile just working out. It isn't strictly
clear that doing this is correct, since someone could somehow rely on
the previous behavior of only being equal when they pointed at the same
object in memory, but UploadedFile is a small wrapper class for BytesIO
with some metadata attributes, so I don't expect people to treat it as
more than an opaque blob or possibly extract some of those attributes
very often, and issues would involve mutating the returned
UploadedFile (since two instances that were previously equal will still
be equal, and this only is misleading if just one is mutated). So it
seems safe to do this.