Make widget_id hashes stable between reruns
Python's native hash function uses a random seed for security purposes,
which means that widget_ids for unkeyed widgets are *not* stable when
streamlit is stopped and restarted.

This doesn't currently make a difference in how anything behaves, but it
did come up when I was investigating a bug and caused a lot of confusion
then, so it seems like something worth fixing.