solve #3604
There's an issue with formerly lineno 365, where streamlit *checks* if the SVG code starts with a `<svg`. This is obviously not the case with most SVG files, and hence it spits out a `FileNotFoundError` when the file starts with anything other than `<svg`. This is solved when the `if` statement that checks for `startswith()` is removed, so the SVG file is rendered regardless of  the starting tag(s)/strings.