Retain file_uploader state when an expander closes (#3660)
Currently, st.file_uploader and st.expander don't work well together as
the file_uploader widget loses its state (only on the frontend) when an
expander is closed and reopened. The reason for this is that we aren't
populating the initial value of a file_uploader widget if there's state
for the widget in the WidgetStateManager.

To fix this, we had to do a decent amount of extra work to send more
info about the file_uploader widget's state back to the server.
Previously, we somewhat hackily encoded all of the information in an
array of ints, where the first element in the array is the ID of the
most recently uploaded file, and the others are the IDs of the active
files. We now send a more descriptive proto back and forth, which
includes the max file ID in one field and a list of protos containing
the id/name/size of the uploaded files in another.

Closes #3557