Coerce values for multiselect/slider/selectbox when they are invalidated
When the arguments to a keyed widget change, it's possible for the value
we have stored in `st.session_state.<widget_key>` to no longer be a
valid value for the widget.

This commit implements a prototype for one way of handling this, where
we

* coerce the value of the widget to something reasonable
* give the user an st.warning telling them what happened and suggesting
  that they `del` the widget's key from `st.session_state`.

Note that we only implement value coercion for the multiselect,
selectbox, and slider widgets since these three widgets are the three
that are most likely to run into the problem we're trying to fix.