Fix "'setIn' cannot be called on an ElementNode" error (#3693)
In the `ReportQueue.enqueue` "Delta composition" logic, we no longer replace `add_block` deltas.

Replacing `add_block` deltas is problematic, because blocks can have other dependent deltas later in the queue. For example:
```python
placeholder = st.empty()
placeholder.columns(1)
placeholder.empty()
```

The call to `placeholder.columns(1)` creates two blocks, a parent container with delta_path `(0, 0)`, and a column child with delta_path `(0, 0, 0)`. If the final `placeholder.empty()` Delta is composed with the parent container Delta, the frontend will throw an error when it tries to add that column child to what is now just an element, and not a block.

(This PR also has type annotations and cleanup for all of `report_queue.py`).

Fixes #3659