Upgrade cypress to version 6.9.1 (#3726)
I previously unsuccessfully attempted to upgrade cypress but didn't spend too
long on it that time around as there was no real incentive to be on a newer
version. Now, we want to be able to download files and inspect them (to test
the download button, naturally), and some features were added starting with
Cypress 6.3 that make this a lot easier to do.

Note that Cypress 8 is the latest major version, but if I'm remembering
correctly, the amount of work required to upgrade to version 7 or higher
is much more than that needed to get to this version, and this version
has the features that we need.