Experiment with explicit dict stack management
Removing the eager conversion of user keys into matching widget ids when
they are known, and so avoiding some bookkeeping.

By making `__getitem__` do the appropriate lookup for any combination of
user key and widget id, lots of callsites are simplified, and combined
with implementing `keys` to do lazy normalization to widget ids, much of
the code that previously needed to know about dict ordering and keys vs
ids can instead be written without it.