Improved Hamburger Menu with changes to Overridden About (#3770)
These code changes will allow for Streamlit to check for whether or not we are in local mode, deployed as a developer mode, or viewer mode.

These changes involve changing the set_page_config api with a new parameter called menu_options. In addition, there are changes to the PageConfig Proto.

On the frontend, there are changes to MainMenu.tsx and StreamlitDialog.tsx (for the about menu).
    - The changes to Mainmenu.tsx involve creating two submenus, style overriding for menuItems, and checking for 
    the keys from s4t and if about is overwritten in order to display correct menu. For StreamlitDialog.tsx, this checks if 
    override is overridden and then properly the overridden message with Streamlit marketing message. 

Finally, there are tests for python changes and typescript changes. E2E tests still need to get the snapshot of about properly written.