(Hopefully) Fix CI, and also suppress some warnings
It remains unknown why CI is fine for new PRs but explodes for the `develop`
pipeline, but it seems like the root cause of the issue is a problem with
Cypress itself that remains unresolved (see https://github.com/cypress-io/cypress/issues/5743
and https://github.com/cypress-io/cypress/issues/7306 for more information).

One of the comments in the issues linked above suggested passing
`{ force: true }` as an argument to the `click` method to get this to
work, and this seems to be a fix on my local machine.

We seem to have run into similar issues before because I remember having
to do the same to get other e2e tests to pass.

I also changed one styled-component tag type from `li` to `ul` to
suppress a warning about invalid tag nestings.