Make calling st.experimental_rerun() within a callback a no-op
Since callbacks are hoisted to run before the actual script run, there's
generally no need to use the st.experimental_rerun() function within a
callback.

Currently, attempting to do so breaks the streamlit app. This PR fixes this so
that calling `st.experimental_rerun()` in a callback is a no-op, and a
warning gets printed to the user's streamlit app to warn the user of
this.

Closes #3854