Make get_module_paths always return absolute paths (#3939)
It turns out that #3732 is caused by the path extractor functions in
get_module_paths sometimes returning relative paths, which confuses
file_util.file_is_in_folder_glob a bit later when we're deciding whether
or not to watch a file. The fix here is to just always convert to
absolute paths. Note that this is a no-op if the path is already an
absolute path, which seems to be the case most of the time.

Closes #3732