Don't hash mtime of tty files and ignore the tell.
I think being able to write to console is helpful when you debug your streamlit application and should not be considered as a side effect that you think about when caching. Otherwise one can easily break a streamlit app by simply changing the logging configuration. This is something that has bitten us on several occasions.