Add PEP 561 support to Streamlit
Streamlit itself already has (some) type annotations, and we type-check the library itself.
But: apps that install Streamlit via pip can not currently typecheck that Streamlit import
because Streamlit is not a "PEP 561 compatible package".

https://mypy.readthedocs.io/en/stable/installed_packages.html

(Note that this does *not* mean we need to type-annotate everything in Streamlit; we can
continue to gradually type as we go.)

(This is a pre-req for addressing https://github.com/streamlit/streamlit/issues/4018)