memo and singleton decorator annotations (#4045)
Adds type annotations to `@st.experimental_memo` and `@st.experimental_singleton`.

This means that running mypy on a Streamlit app should catch incorrect usage of a cached function, e.g.

```python
import streamlit as st

@st.experimental_memo
def foo(value: int) -> None:
  st.write(value * 2)

foo("bar")  # this is now a mypy error, whereas before it was not
```

This changes the signatures for memo + singleton _ever so slightly_, in that it makes all "decorator arguments" into keyword-only args. It was previously legal to pass them positionally. I think this is ok because (1) we're still experimental and (2) in practice it's unlikely people were passing these parameters positionally anyway.