Ignore type errors introduced by mypy 0.920
Previously, these functions returning sets was considered valid by mypy
and the python docs, but now they are expected to be actual view
objects. Unfortunately the constructors for the view objects assume your
class is a dumb wrapper around a dict, and so don't work correctly with
`WStates` for items/values (because it relies on you calling
`__getitem__`), and also don't work well with `SessionState` merging and
key translating several dicts. So we will mostly ignore this, except for
keys view for `WStates`, which should work fine as a `KeysView`.