Wait for enough elements to load in flaky e2e tests (#4193)
The core problem is described in
https://docs.cypress.io/guides/core-concepts/retry-ability#Only-the-last-command-is-retried

Chaining `get` and then immediately `eq` is prone to flakiness, because
if the get call sees some but not enough elements, the eq call will
fail, but only the eq call will be retried, so it'll continue to fail
even when the actual page loads the elements before the eq call times
out. Adding an assertion that there are enough elements ensures that the
initial query will be retried until enough elements are present.