Add length checking before indexing in every e2e
This is a sledgehammer. Not all of the places changed were vulnerable to
flakiness, as I did not have a convenient way to exclude those from my
semgrep check.

This does not address indexing using `.last()`, which requires I figure
out how many elements are actually expected. Possibly it should be
replaced by an explicit index.

This does not address other ways of querying (like `.find()`) that we
then index into, which are probably also vulnerable to flakiness.

Probably we should see how difficult it is to write and use a
checked-indexing function, which would cut down on the repetiton.