Get rid of preheated script runs (#4259)
* Get rid of preheated script runs

When a streamlit server is first started, we currently trigger a run of the
script defining an app and save the resulting deltas so that the very first
page load of an app can be more or less instantaneous.

This optimization is currently not too helpful given how streamlit is used in
practice today (it was originally added to make long-running jobs started
via `streamlit run` feel fast, but people generally don't use streamlit to kick
off long-running computations). Furthermore, we'll soon be adding some features
that won't play nicely with the optimization. In particular, the upcoming
`st.user` feature interacts with script preheats weirdly as the information
required to populate `st.user` doesn't exist in a preheat run.

Given complications that will arise in the near-future as well as the fact that
the optimization itself is a vestigial one, it seems like it's time to remove
preheated script runs.

* Rework cli_smoke_tests to no longer rely on script preheats

* Try making tests less timing sensitive

* Revert an unintended change in an e2e test script

* Replace `%` usage with an f-string