Make checkbox/radio only show focus when tabbbing, not when clicked.
Focus queues are often confusing for mouse-based users (why is there
sometimes a box around this element?). To help with this, in this PR I
replace the "focus" state with the "focus-visible" state, which tries to
be smart about when to draw focus queues around an element.

Here's a nice little write-up about "focus-visible":
https://matthiasott.com/notes/focus-visible-is-here