Get version using importlib-metadata instead of pkg_resources
`pkg_resources` is pretty slow, adding up to .75s to our startup time.
Importlib is much faster, and the `version` function from
`importlib.metadata` gives us exactly what we want.

We still support 3.7, so we have to use the separate package that
provides a backport, because the `metadata` module wasn't added to
stdlib importlib until 3.8.