Reduce websocket message handling latency (#4355)
By default, a JavaScript websocket delivers its data in the `Blob` format, and we need to use a FileReader to asynchronously convert to an `ArrayBuffer`. But we can just ask for ArrayBuffer data up front, and skip this FileReader step. On my Mac, running an internal Streamlit dashboard app, this saves us ~5-50 milliseconds of latency per websocket message.