Small refactor: session_data no longer stores "script_run_id" (#4379)
`SessionData` is intended to store session-wide data, but it also currently holds `script_run_id`: a random ID that's regenerated every time a new script run starts.

This doesn't make tons of sense, especially because the ID is only used in a single place: the function where it's regenerated. It'll also be problematic if we move to a world where a single session can have multiple active ScriptRunners.

This PR just fixes this!