Disable Websocket compression by default (#4363)
WS compression saves a fair bit of network data, but compressing large
dataframes takes a very long time (6.5s for a random 150MB DF), which
destroys the responsiveness of apps using them. For slower network
connections it is unclear if this is worthwhile, but for fast ones it
appears much better to have compression off by default.

Using smarter heuristics for when to enable compression is worth
exploring in the future.