RerunAnalyzer (#4357)
Super-basic recording and analysis of performance-related events that happen during a rerun

- `PerformanceEvents` is a simple static class that records typed, timestamped events
- `App` and `WebsocketConnection` log these events
- `RerunAnalyzer` generates a simple JSON report of the performance analysis for a single rerun

This may or may not be code that we end up shipping in Streamlit. If we do ship it, it'll need tests and maybe better documentation and a better way to enable and disable "performance logging" mode. But in the meantime, it'll be useful for measuring the impact of our "make streamlit reruns faster" work.