SessionState widget registration refactor (#4432)
Widget registration currently involves calls to multiple SessionState functions:
- `set_keyed_widget_metadata`
- `set_unkeyed_widget_metadata`
- `should_set_frontend_state_value`
- `get_value_for_registration`

This PR collapses that down to a single public API, `SessionState.register_widget`. The idea is to reduce SessionState's public API surface area ahead of upcoming changes in the `feature/faster-reruns` branch.

There's a wee bit of other cleanup in here as well that is called out in comments below.