ScriptRunner.on_event: include 'sender' param (#4436)
Modify `ScriptRunner.on_event` to include the sender of the event in the call. (This is actually the way that signals are intended to be used - we'd been sending the event data in the signal's "sender" param.)

This will become important when there can be multiple ScriptRunners executing inside a single AppSession.