AppSession: handle script events on the main thread (#4467)
Refactors AppSession to handle all ScriptRunner events on the main thread.

- ScriptRunner no longer takes an `enqueue` callback param. Instead, it signals that it wants to enqueue a new ForwardMsg via the `ScriptRunnerEvent.ENQUEUE_FORWARD_MSG` event. (This means that there's now only one channel that ScriptRunner uses to communicate back to AppSession - the `on_event` signal.)
- `AppSession.enqueue` is now a private function called `AppSession._enqueue_forward_msg`, for clarity
- `AppSession._on_scriptrunner_event` performs all its logic in an ioloop callback, which means that all ScriptRunner events are handled on the main thread _only_. (Previously, _some_ scriptrunner events were handled in a callback, and others were handled in line.)
- `app_session_test.py` now has a test for the very convoluted `handle_backmsg_exception` function.  (This function should be targeted for refactor down the line!)
- There's also a small amount of unrelated cleanups in `app_session_test.py`

This is one of the last big pieces that needs to be in place for the "faster-reruns" branch. With this change, AppSession will be able to safely identify and ignore ScriptRunner events that come from a "zombie" ScriptRunner that's still executing. 

It also (hopefully) makes AppSession easier to reason about: with the exception of the two functions that explicitly enqueue ioloop callbacks to be performed on the main thread, AppSession now executes entirely on the main thread!