Remove mostly useless speedup in ping retries
This logic was supposed to have the ping timer progress even as we
waited for the ping request to resolve, but it took a milisecond
duration and converted it into seconds, then subtracted it from a
milisecond value, so it did almost nothing in practice. So it won't hurt
anything to remove to simplify the function.