Cache LocalSourcesWatcher modules
The modules that have been imported by the python interpreter will
usually not change between script runs, and the filesystem lookups to
get the modules' paths are expensive, so we want to skip this when
possible.

We use the set of module names (the keys of sys.modules) to determine if
we need to regenerate the paths. This could miss an already imported
module being replaced by another version of itself with a different file
path, but that's quite an edge case, and it would only miss out on us
being able to see changes to the module's file to prompt the user to
rerun the script, which seems fine.