Prevent SessionState mutation in stopped ScriptRunner (#4624)
### Motivation

An app using the faster-reruns option may spin up a new ScriptRunner to handle a rerun request even as its current ScriptRunner is in the process of stopping. When this happens, we want to avoid race conditions that might occur when both script executions are accessing the same SessionState data simultaneously.

### SafeSessionState

This PR introduces `SafeSessionState`, which is a thread-safe wrapper around `SessionState`. It has a `disconnect` function that's called by ScriptRunner when it gets a stop request. A "disconnected" SafeSessionState no longer reads and writes its underlying SessionState instance - all operations becomes no-ops (or raise KeyErrors, as appropriate).

This means that a script running inside a ScriptRunner that is shutting down will not mutate session state (or register widgets).

(I don't love the name "SafeSessionState", but I couldn't think of a better one! "DisconnectableSessionState" is maybe more accurate, but is too much of a mouthful?)

### Other minor bits

- `SessionState.clear_state()` -> `SessionState.clear()`
- `SessionState.as_widget_states()` -> `SessionState.get_widget_states()`
- `SessionState._keys()` is now private
- `validate_key()` -> `require_is_valid_user_key()`