BrowserWebSocketHandler.on_message: fix type signature (#4896)
`WebSocketHandler.on_message` actually takes a `Union[str, bytes]`. We only handle `bytes` in our subclass, so raise an exception if the frontend incorrectly sends us a str.

(The protobuf `ParseFromString()` function will throw an error if it receives a `str`, so this doesn't actually introduce a new exception path into the lib, it just makes the exception that will inevitable be thrown a bit more understandable.)

The `@tornado.gen.coroutine` decorator strips our type signature, which is why this isn't caught by mypy. We're going to ditch this decorator in the near future.