Add type annotations for register_widget (#4813)
## 📚 Context

I was looking into improving some type annotations for widgets. One major blocker I encountered was the fact that the `register_widget` function did not have useful return types. The return type should really be dependent on the serializer and deserializer callable you pass in. This PR addresses that by introducing appropriate type variables for the serializer and deserializer, such that, if they are properly annotated, the return value from `register_widget` will also have a narrowed type.

One issue I discovered while working on this is that many widgets does not seem to handle the case where a disconnect occurs, and the widget value returned from `register_widget` is None. I tried to handle this by unifying the fallback logic, such that the disconnect case is handled the same way as when `ctx` is missing. I don't see a reason why these were different in the first place.

- What kind of change does this PR introduce?
  - [X] Bug fix: Unified the fallback logic when the true widget state could not be determined 
  - [X] Other, please describe: Type annotations

---

**Contribution License Agreement**

By submitting this pull request you agree that all contributions to this project are made under the Apache 2.0 license.