AppSession.handle_backmsg (#4982)
Moves `BackMsg.type` switching logic out of `BrowserWebSocketHandler` and into `AppSession`. This logic is not SessionClient-specific; moving it into AppSession means that other future SessionClient implementations don't have to duplicate it.

This is just a refactor, not a logic change.

We don't have much in the way of AppSession message handling tests - this PR doesn't do much to address that, but it does add a single new test that asserts that exceptions in `AppSession.handle_backmsg` get handled internally.