SessionClientDisconnectedError (#4985)
Currently, `Server. _loop_coroutine` catches `tornado.websocket.WebSocketClosedError` errors to detect when a client is disconnected. This ties `Server` to Tornado WebSocket-based `SessionClient` implementations, which we don't want.

This PR introduces `SessionClientDisconnectedError`, which SessionClient instances should raise when they're disconnected. BrowserWebSocketHandler re-raises any WebSocketClosedError as a SessionClientDisconnectedError.

This also adds a test to ensure that graceful session cleanup happens when a client raises an error.