Don't expose SessionInfo to UploadFileRequestHandler (#4997)
Currently, `UploadFileRequestHandler` takes a reference to server's private `_get_session_info` function, which returns SessionInfo instances. But SessionInfo is really a private implementation detail of server, so we'd rather not expose it externally.

This PR adds a `Server.is_active_session()` function that doesn't expose SessionInfo, so that UploadFileRequestHandler is decoupled from SessionInfo.