Don't expose SessionData class (#5071)
`SessionData` is an internal implementation detail of `AppSession` (it contains the `ForwardMsgQueue` that we use to send forward messages to the frontend). 

We're currently unnecessarily exposing it to `ScriptRunner` and `LocalSourcesWatcher`; these classes actually only need `SessionData.main_script_path`.

This PR removes the coupling, but doesn't change any functionality. (This is part of a larger upcoming "streamlit.runtime" refactor.)