passing coroutines to asyncio.wait is deprecated - pass tasks instead (#5129)
Passing coroutines directly `asyncio.wait` [is deprecated](https://docs.python.org/3/library/asyncio-task.html#asyncio.wait) and will raise an error in Python 3.11.

The proper thing to do here is to wrap coroutines in Tasks (which `asyncio.wait` does under the hood). This PR does that (and also passes the tasks as a tuple instead of a list as a micro-optimization).