AppSession: allow running on another thread (#5196)
Currently, `AppSession._handle_scriptrunner_event_on_main_thread` asserts that it's called on the main thread. This is subtly incorrect - the function must be called on the AppSession's asyncio event loop thread, which is generally but not necessarily the main thread.

(Specifically, Snowpark is currently running the Streamlit Runtime on another thread.)

This PR changes the assertion - we now assert that the function is called on the event loop. (The PR also renames the function.)