MediaFileManager.add: just return URL (#5227)
`MediaFileManager` has a `MediaFile` type that's currently being returned from MediaFileManager.add. 

Ultimately, `MediaFile` should be internal - it will be an implementation detail instead of part of the public API. This PR is a step in that direction - `MediaFileManager.add` now just returns the URL of the added file, rather than the `MediaFile` object itself.