Convert widget serde function closures to objects (#5226)
As part of widget replay, I want to store `WidgetMetadata` in cache
results, but memo requires everything be pickleable, and functions
defined not at the top level can't be pickled. So I used the equivalence
between closures and objects, and converted the widget serde functions
into methods on a top level object.