Remove emoji support from `image_to_url` (#5270)
`images.image_to_url` currently takes an `allow_emoji` param; if the input string is an emoji and this flag is True, we return the input string as-is (despite the input string not being a URL). This flag is only used by the `set_page_config` command, which allows for setting emoji as page icons.

This PR removes the emoji bits from `image_to_url` and moves them to an internal `_get_favicon_string` function in `page_config.py`. It also adds additional test coverage for `page_config` icon-related stuff.

This emoji-removal makes the upcoming MediaFileManager refactor easier to reason about! (And also makes image_to_url actually work as advertised.)