st_set_page_config_icon e2e test: use a favicon in a known-good location (#5288)
`st_set_page_config_icon` walks our entire directory structure to search for a specific icon file. This means the script is spelunking into e.g. `node_modules`, which is an endless pit of despair; the script takes upwards of 20s to run on my (very fast) machine.

Instead, let's just stick the icon in a known location and reference that path directly. This makes the test complete instantly.