Allow widget keys to generate when widgets are missing
Otherwise there may be exceptions from lookup failures when reconnecting
to a running app. We could instead catch the exception somewhere and
treat it as a cache miss, but this way seems more robust, because it
handles a conditionally created widget that is not currently being used
by allowing caching to still happen based on the other widgets.