MediaFileManager thread-safety (#5332)
MediaFileManager is called from multiple threads, so add a lock.

This doesn't change any behavior, and doesn't introduce any new tests. But I'd like to get more serious about testing our multi-threaded classes in the (near) future, and I've added a task to my list to figure out a plan for this.