metrics_util_test.py: fix patch statements (#5627)
Drive-by PR to fix a couple `patch` statements that are implemented incorrectly. 

(The intention is to have `os.path.isfile` return True for a specific path, but as written the patches would return True for all paths. This didn't break the tests - it was just a bit confusing as a reader.)