Plotly Customization (#5681)
Right now, the current implementation of st.plotly_chart(...,theme="streamlit") doesn't allow overriding properties within a graph; it will always applies streamlit theme to that graph. As a result, it's not very usable for a developer who wants to override the colors or other such specific things but also get Streamlit's layout changes. As a result, this PR is for that.
This is mainly refactoring and allowing theme="streamlit" to be the default. We are using Plotly's inbuilt theme object and then passing that to the frontend because we want to swap the colors when the settings are swapped from light to dark and the backend has no knowledge of that.

Demo app: https://streamlit-feature-demos-plotlystreamlitstreamlit-app-jjumlv.streamlit.app