Hoist element message data creation out of loop
We only want to reset the callstack's local data after we are done
adding the widget/media element to the message list for each cached
function we are running in. Moving the creation of the message data
class out of the loop, and saving it to each of the lists, accomplishes
this with the bonus of saving some memory use.