Refactor AppSession to simplify improving reconnect behavior (#5782)
In the `feature/session-manager` branch, we're working towards improving the way that we handle
websocket disconnects. As part of this work, we'll need to have the final steps of `AppSession`
cleanup (done by the `shutdown()` method) happen when a session is garbage collected rather
than when the browser's websocket connection is dropped.

To make this easier, this PR splits out file changed handler registration/deregistration into two helper
methods. We also add a `__del__` method to `AppSession`, which calls the `shutdown()` method on
garbage collection.

Finally, we add a test to make it harder to introduce `AppSession`-related memory leaks related to file
watchers holding references to `AppSession` file changed handler methods. This exact issue has
caused memory leaks in the past, so we should be extra-careful to avoid it given that our work to
defer `AppSession` cleanup is particularly prone to running into this.