Replace closeConnection with shutdownRuntime and disconnectWebsocket (#5809)
This is another change that's being made because we need it for the feature/session-manager branch
but is being made straight to develop to keep the final diff size down.

The special test-only closeConnection proto message type has a pretty misleading name because what
it really does is shut down the runtime entirely. Because we need a way to test reconnect behavior (and thus
a way to close only the browser websocket connection without nuking the entire Streamlit server), we replace
this proto type with debugShutdownRuntime and debugDisconnectWebsocket, which do what their names
suggest.